TRUS-3178 Rename Connectors
