+++ app/connectors/DesConnector.scala
@@ -21,12 +21,8 @@ import java.util.UUID
 import javax.inject.Inject
 import play.api.Logging
 import play.api.http.HeaderNames
-import play.api.libs.json._
 import config.AppConfig
 import models._
-import models.getEstate.GetEstateResponse
-import models.variation.VariationResponse
-import services.EstatesStoreService
 import utils.Constants._
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.http.HttpClient
@@ -34,22 +30,9 @@ import uk.gov.hmrc.http.HttpClient
 import scala.concurrent.ExecutionContext.Implicits._
 import scala.concurrent.Future
 
-class DesConnector @Inject()(http: HttpClient, config: AppConfig, estatesStoreService: EstatesStoreService) extends Logging {
+class DesConnector @Inject()(http: HttpClient, config: AppConfig) extends Logging {
 
-  private lazy val estatesServiceUrl : String = s"${config.estatesBaseUrl}/estates"
-
-  private lazy val matchEstatesEndpoint : String = s"$estatesServiceUrl/match"
-
-  private lazy val estateRegistrationEndpoint : String = s"$estatesServiceUrl/registration"
-
-  // When reading estates from DES, it's the same endpoint as for trusts.
-  // So this must remain "trusts" even though we're reading an estate.
-  private lazy val getEstateUrl: String =  s"${config.getEstateBaseUrl}/trusts"
-
-  private def create4MLDEstateEndpointForUtr(utr: String): String = s"$getEstateUrl/registration/$utr"
-  private def create5MLDEstateEndpointForUtr(utr: String): String = s"$getEstateUrl/registration/UTR/$utr"
-
-  private lazy val estateVariationsEndpoint : String = s"${config.varyEstateBaseUrl}/estates/variation"
+  private lazy val subscriptionsUrl : String = s"${config.desEstatesBaseUrl}/trusts"
 
   private val ENVIRONMENT_HEADER = "Environment"
   private val CORRELATION_HEADER = "CorrelationId"
@@ -62,50 +45,13 @@ class DesConnector @Inject()(http: HttpClient, config: AppConfig, estatesStoreSe
       CORRELATION_HEADER -> correlationId
     )
 
-  def checkExistingEstate(existingEstateCheckRequest: ExistingCheckRequest)
-  : Future[ExistingCheckResponse] = {
+  def getSubscriptionId(trn: String): Future[SubscriptionIdResponse] = {
+
     val correlationId = UUID.randomUUID().toString
 
     implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
 
-    logger.info(s"[checkExistingEstate] matching estate for correlationId: $correlationId")
-
-    http.POST[JsValue, ExistingCheckResponse](matchEstatesEndpoint, Json.toJson(existingEstateCheckRequest))
-  }
-
-  def registerEstate(registration: EstateRegistration): Future[RegistrationResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.info(s"[registerEstate] registering estate for correlationId: $correlationId")
-
-    http.POST[JsValue, RegistrationResponse](estateRegistrationEndpoint, Json.toJson(registration)(EstateRegistration.estateRegistrationWriteToDes))
-  }
-
-  def getEstateInfo(utr: String): Future[GetEstateResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc : HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.info(s"[getEstateInfo][UTR: $utr] getting playback for estate for correlationId: $correlationId")
-
-    estatesStoreService.is5mldEnabled.flatMap { is5MLD =>
-      if (is5MLD) {
-        http.GET[GetEstateResponse](create5MLDEstateEndpointForUtr(utr))(GetEstateResponse.httpReads(utr), implicitly[HeaderCarrier](hc), global)
-      } else {
-        http.GET[GetEstateResponse](create4MLDEstateEndpointForUtr(utr))(GetEstateResponse.httpReads(utr), implicitly[HeaderCarrier](hc), global)
-      }
-    }
-  }
-
-  def estateVariation(estateVariations: JsValue): Future[VariationResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.info(s"[estateVariation] submitting estate variation for correlationId: $correlationId")
-
-    http.POST[JsValue, VariationResponse](estateVariationsEndpoint, Json.toJson(estateVariations))
+    val subscriptionIdEndpointUrl = s"$subscriptionsUrl/trn/$trn/subscription"
+    http.GET[SubscriptionIdResponse](subscriptionIdEndpointUrl)
   }
 }
