+++ conf/application.conf
@@ -147,14 +147,6 @@ microservice {
       port = 8500               # AUTH
     }
 
-    des-estates-non-migrated {
-        protocol = http
-        host = localhost
-        port = 9805             # ESTATE_REGISTRATION_STUB
-        token = devToken        # Shared token for register, playback and variation
-        environment = dev       # Shared env for register, playback and variation
-    }
-
     des-estates {
         protocol = http
         host = localhost
@@ -163,13 +155,21 @@ microservice {
         environment = dev       # Shared env for register, playback and variation
     }
 
-    des-estates-playback {
+    ifs-estates {
+        protocol = http
+        host = localhost
+        port = 9805             # ESTATE_REGISTRATION_STUB
+        token = devToken        # Shared token for register, playback and variation
+        environment = dev       # Shared env for register, playback and variation
+    }
+
+    ifs-estates-playback {
         protocol = http
         host = localhost
         port = 8833             # ESTATES_PLAYBACK_STUB
     }
 
-    des-estates-variation {
+    ifs-estates-variation {
         protocol = http
         host = localhost
         port = 8834             # ESTATES_VARIATIONS_STUB
