+++ it/transforms/variations/AmendPersonalRepSpec.scala
@@ -28,7 +28,7 @@ import play.api.libs.json.{JsValue, Json}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import uk.gov.hmrc.auth.core.AffinityGroup.Organisation
-import connectors.DesNonMigratedConnector
+import connectors.DesConnector
 import controllers.actions.{FakeIdentifierAction, IdentifierAction}
 import models.getEstate.GetEstateResponse
 import models.variation.{EstatePerRepIndType, PersonalRepresentativeType}
@@ -45,7 +45,7 @@ class AmendPersonalRepSpec extends AsyncFreeSpec with MustMatchers with MockitoS
 
   "an amend personal rep call" - {
 
-    val stubbedDesConnector = mock[DesNonMigratedConnector]
+    val stubbedDesConnector = mock[DesConnector]
     when(stubbedDesConnector.getEstateInfo(any())).thenReturn(Future.successful(getEstateResponseFromDES))
 
     val cc = stubControllerComponents()
@@ -59,7 +59,7 @@ class AmendPersonalRepSpec extends AsyncFreeSpec with MustMatchers with MockitoS
       ): _*)
       .overrides(
         bind[IdentifierAction].toInstance(new FakeIdentifierAction(cc.parsers.default, Organisation)(ExecutionContext.global)),
-        bind[DesNonMigratedConnector].toInstance(stubbedDesConnector)
+        bind[DesConnector].toInstance(stubbedDesConnector)
       )
       .build()
